work in progress

+++ SqlAnalytics.Domain/Models/SqlPlanStatisticsModel.cs
@@ -16,10 +16,10 @@ namespace SqlAnalyticsManager.Models
         public string DatabaseName { get; set; }
         public string SchemaName { get; set; }
         public string TableName { get; set; }
-        public decimal EstimateIO { get; set; }
-        public decimal EstimateCPU { get; set; }
-        public decimal EstimateRows { get; set; }
-        public decimal EstimateTotalSubTreeCost { get; set; }
+        public double EstimateIO { get; set; }
+        public double EstimateCPU { get; set; }
+        public double EstimateRows { get; set; }
+        public double EstimateTotalSubTreeCost { get; set; }
         public string StatementText { get; set; }
         public int NodeId { get; set; }
         public int ParentNodeId { get; set; }
